# optimize norm on ATen CPU backend