@ezyang 2018-09-12
@pytorchbot retest this please