@mingfeima 2018-09-12
@ezyang please help review the fails. Seems not related to this PR.