@fmassa 2018-09-12
Looks like this PR is a duplicate of https://github.com/pytorch/pytorch/pull/10535