@ezyang 2018-09-12
The failures are unrelated. @mingfeima is it safe to assume you resubmitted this PR on @xhzhao's behalf?