@ezyang 2018-09-13
For my part, the patch looks basically reasonable, but I am not an expert in the AVX parallelization.