@cpuhrsch 2018-09-13
Generally looks good - @mingfeima I'd also be curious about and much appreciative of your opinion on the user-experience of using the abstractions (vec256 and parallel) and any feedback related to it, if you want to give any.