@mingfeima 2018-09-13
@fmassa @ezyang this one is authored by @xhzhao, we are on the same team. Pretty much you see the entire intel-pytorch team on this thread. Previously we also have @MlWoo on the team, yet he quits this summer.
I thought xhzhao hadn't pull request it... my mistake... anyway, either one should be OK.