@mingfeima 2018-09-13
@cpuhrsch thanks for the review, will do.
Honestly i'd like to say `parallel` and `vec256` work like a charm. With `parallel` and `vec256` it is very easy to construct high performance kernels in `ATen/native`, without touching any `pragma omp` and intrinsic.
A little suggestion is that you probably need to add `omp_in_parallel` check inside parallel_for and parallel_reduce. It could be possible that parallel is called in nested loop unintentionally, in that case, it's going to cause oversubscription.